+++ src/helpers/partition.js
