+++ src/http/DataProduceEndpoints.js
@@ -4,7 +4,7 @@ const { StreamMessage } = require('streamr-client-protocol').MessageLayer
 const { InvalidJsonError } = require('streamr-client-protocol').Errors
 
 const FailedToPublishError = require('../errors/FailedToPublishError')
-const partition = require('../partition')
+const partition = require('../helpers/partition')
 
 const authenticationMiddleware = require('./RequestAuthenticatorMiddleware')
